Add property-based tests for FlowLog engine and fix sum overflow
Add 46 proptest cases across row types, Min semiring, comparisons,
arithmetic, aggregation, and stratification. Fix integer overflow in
sum aggregation by accumulating in i128 before truncating to i64.
Migrate internal value storage from i32 to i64 throughout the codebase.

+++ dbflow-core/src/compiler/emit.rs
@@ -98,9 +98,9 @@ pub fn emit_datalog(result: &CompileResult) -> String {
 }
 
 /// Write EDB facts to `.facts` files in the given directory.
-/// Each file is named `{relation_name}.facts` with tab-separated i32 values.
+/// Each file is named `{relation_name}.facts` with tab-separated i64 values.
 pub fn write_facts(
-    edb_facts: &HashMap<String, Vec<Vec<i32>>>,
+    edb_facts: &HashMap<String, Vec<Vec<i64>>>,
     output_dir: &Path,
 ) -> Result<(), CompileError> {
     fs::create_dir_all(output_dir)?;
