+++ flowlog/src/macros/src/lib.rs
@@ -417,7 +417,7 @@ pub fn codegen_min_optimize(_: TokenStream) -> TokenStream {
                         for i in 0..#key_arity {
                             key.push(row.column(i));
                         }
-                        let value = row.column(#key_arity) as u32;
+                        let value = row.column(#key_arity) as u64;
                         std::iter::once((key, reading::Min::new(value))).into_iter().map(move |(x, d2)| (x, t.clone(), d2))
                     })
                     .as_collection()
@@ -450,7 +450,7 @@ pub fn codegen_min_optimize(_: TokenStream) -> TokenStream {
                             result.push(key.column(i));
                         }
                         // Push minimized value into the last column (extracted from diff!)
-                        result.push(min_val.value as i32);
+                        result.push(min_val.value as i64);
                         std::iter::once((result, reading::semiring_one())).into_iter().map(move |(x2, d2)| (x2, t.clone(), d2))
                     })
                     .as_collection()
@@ -477,7 +477,7 @@ pub fn codegen_min_optimize(_: TokenStream) -> TokenStream {
                         }
 
                         // Extract the last column as the value to minimize
-                        let value = row.column(arity - 1) as u32;
+                        let value = row.column(arity - 1) as u64;
                         std::iter::once((key, reading::Min::new(value))).into_iter().map(move |(x, d2)| (x, t.clone(), d2))
                     })
                     .as_collection()
@@ -510,7 +510,7 @@ pub fn codegen_min_optimize(_: TokenStream) -> TokenStream {
                         }
 
                         // Push minimized value as the last column
-                        result.push(min_val.value as i32);
+                        result.push(min_val.value as i64);
 
                         std::iter::once((result, reading::semiring_one())).into_iter().map(move |(x2, d2)| (x2, t.clone(), d2))
                     })
